+++ industrial_sem3/pain/src/BaseGUILogic/TextBox/TextBoxDelegate.hpp
@@ -6,7 +6,7 @@ class TextBoxDelegate
 public:
     virtual ~TextBoxDelegate() {};
 
-    virtual void onTextChange(const char *newText, int userData) = 0;
+    virtual void onTextChange(const char *newText, uint64_t userData) = 0;
 };
 
 #endif
