+++ industrial_sem3/pain/src/BaseGUILogic/Window/WindowCloseDelegate.hpp
@@ -10,7 +10,7 @@ public:
     WindowCloseDelegate() = delete;
     WindowCloseDelegate(Window *window) : mWindow(window) {};
 
-    virtual void onClick(int userData) override
+    virtual void onClick(uint64_t userData) override
     {
         mWindow->scheduleForDeletion();
     }
