+++ industrial_sem3/pain/src/EditorLogic/PluginPickerWindow/PluginPickerCloseDelegate.hpp
@@ -12,7 +12,7 @@ public:
     PluginPickerCloseDelegate(PaintController *controller, PluginPickerWindow *window) :
         mWindow(window), mController(controller) {};
 
-    virtual void onClick(int userData) override
+    virtual void onClick(uint64_t userData) override
     {
         mWindow->scheduleForDeletion();
         mController->onPluginPickerClose(mWindow);
