+++ industrial_sem3/pain/src/EditorLogic/SizePickerWindow/SizePickerCloseDelegate.hpp
@@ -12,7 +12,7 @@ public:
     SizePickerCloseDelegate(PaintController *controller, SizePickerWindow *window) :
         mWindow(window), mController(controller) {};
 
-    virtual void onClick(int userData) override
+    virtual void onClick(uint64_t userData) override
     {
         mWindow->scheduleForDeletion();
         mController->onSizePickerClose();
